Added functionality for users to view and delete their registrations

+++ app/Http/Controllers/RegistrationController.php
@@ -23,7 +23,14 @@ class RegistrationController extends Controller
      */
     public function index()
     {
-        return view('registration/index');
+        $children = Child::get()->where('f_id','=', Auth::user()->id);
+        $registrations = DB::table('child')
+        ->join('registration','registration.c_id','child.c_id')
+        ->join('family','family.f_id','child.f_id')
+        ->join('session','session.s_id','registration.s_id')
+        ->select('child.*','session.*','registration.*')
+        ->get();
+        return view('registration/index', compact('children','registrations'));
     }
 
     /**
@@ -36,7 +43,7 @@ class RegistrationController extends Controller
         if (Auth::check()){
             $session = Session::findOrFail($s_id);
             $children = Child::join('family','child.f_id','=','family.f_id')
-                ->get();
+            ->get();
             if(Auth::user()->id == 1)
                 return view('registration.createAsAdmin', compact('session', 'children'));
 
@@ -48,7 +55,7 @@ class RegistrationController extends Controller
         else return redirect()->route('login');
     }
 
-   
+
     public function store(Request $request)
     {
         $attributes = request()->validate([
@@ -61,7 +68,7 @@ class RegistrationController extends Controller
         return $this->checkAgeConstraints($children, $request);
 
 
-       
+
     }
 
 
@@ -78,13 +85,13 @@ class RegistrationController extends Controller
     
 
     public function checkAgeConstraints($children, Request $request){
-        
+
         // retrieve session and its classList
         $r_id = -1;
         $session = Session::where('s_id',request('s_id'))->first();
         $classList = Registration::where('s_id', request('s_id'))
-            ->select('c_id')
-            ->get();
+        ->select('c_id')
+        ->get();
 
         $list=[];
         $success = 'Successfully registered';
@@ -96,9 +103,9 @@ class RegistrationController extends Controller
 
         // dummy validator, to be used later
         $errors = Validator::make($session->toArray(),
-                    ['min_age' => 'required']          
-                ); 
-       
+            ['min_age' => 'required']          
+        ); 
+
         foreach($children as $child){
 
             // get the c_id passed through the request. Store as id
@@ -117,7 +124,7 @@ class RegistrationController extends Controller
                 // if session is full, throw error tht session is full
                 if($session->is_full == 't'){
                     $validator->errors()->add('dum','dum');
-                   $errors->errors()->add('Full',"This session is now full. $child->child_name could not be registered");      
+                    $errors->errors()->add('Full',"This session is now full. $child->child_name could not be registered");      
                 }
 
                 // if c_id is in class list already, throw error that they're already registered
@@ -131,26 +138,26 @@ class RegistrationController extends Controller
                     if($validator->messages()->get('age'))
                         $errors->errors()->add('Age',"$child->child_name could not be registered because (s)he is not within the age limits for this session ($session->min_age to $session->max_age)");
                 }           
-                      
+
                 // if no validation failures, register child  
                 else {
                     $registration = Registration::create([
                         's_id'=>request('s_id'),
                         'c_id'=>$child->c_id,
-                     ]);
+                    ]);
 
                     // add child's name to success message
                     $success = $success . ", $child->child_name";
 
                     $r_id = $registration->r_id;
                     $num_registered = count(Registration::where('s_id',$session->s_id)->get());
-                        
+
                     // update isFull status for session
                     if ($num_registered >= $session->max_attendance){
                         $session->is_full='t';
                         $session->save();
                     }
-                
+
                 }
             }
         }
@@ -170,19 +177,19 @@ class RegistrationController extends Controller
         if ($id == 1){
             $children = DB::table('child')
 
-                ->join('registration','child.c_id','=','registration.c_id')
-                ->where('registration.s_id','=',$registration->s_id)
-                ->select('child.child_name','registration.r_id')
-                ->get();
+            ->join('registration','child.c_id','=','registration.c_id')
+            ->where('registration.s_id','=',$registration->s_id)
+            ->select('child.child_name','registration.r_id')
+            ->get();
         }
         else{
             $children = DB::table('child')
-                
-                ->join('registration','child.c_id','=','registration.c_id')
-                ->where('registration.s_id','=',$registration->s_id)
-                ->where('child.f_id','=', $id)
-                ->select('child.child_name','registration.r_id')
-                ->get();
+
+            ->join('registration','child.c_id','=','registration.c_id')
+            ->where('registration.s_id','=',$registration->s_id)
+            ->where('child.f_id','=', $id)
+            ->select('child.child_name','registration.r_id')
+            ->get();
         }
         $session = Session::find($registration->s_id);
 
@@ -192,7 +199,9 @@ class RegistrationController extends Controller
     
     public function edit(Registration $registration)
     {
-        //
+        $session = Session::find($registration->s_id);
+        $child = Child::find($registration->c_id);
+        return view('registration.edit', compact('session','registration','child'));
     }
 
     
@@ -202,11 +211,30 @@ class RegistrationController extends Controller
     }
 
     
-    public function destroy(Registration $registration)
+    public function destroy(Registration $registration, Request $request)
     {
-       $registration->delete();
-       $session=Session::where('s_id', $registration->s_id)
-            ->update(['is_full'=>'f']);
-        return redirect()->route('session.show',[$registration->s_id]);
-    }
+
+        $child = Child::find($registration->c_id);
+        $session = Session::find($registration->s_id);
+
+        if(Auth::user()->id == 1){
+         $registration->delete();
+         $session=Session::where('s_id', $registration->s_id)
+         ->update(['is_full'=>'f']);
+         return redirect()->route('registration.index');
+        }
+        elseif ($session->date >= date('Y-m-d', time() - 60*60*24)) {
+            $errors = 'We do not accept online cancellations within 24 hours of the session date. Please call Scout Island Nature Centre at 250-398-8532 to cancel. You will not be refunded.';
+            $request->session()->flash('errors', $errors);
+            return redirect("/registration/$registration->r_id/edit");
+        }
+        else {
+            $success = "Successfully cancelled registration for $child->child_name on $session->date at $session->start_time";
+            $registration->delete();
+            $session=Session::where('s_id', $registration->s_id)
+                ->update(['is_full'=>'f']);
+            $request->session()->flash('success', $success);
+            return redirect('/registration');
+        }
+     }
 }
